Move NowContainer
