+++ app/assets/javascripts/class_lists/ClassListCreatorWorkflow.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
-import {withDefaultNowContext} from '../../../../spec/javascripts/support/NowContainer';
 import _ from 'lodash';
+import {withDefaultNowContext} from '../testing/NowContainer';
 import ClassListCreatorWorkflow from './ClassListCreatorWorkflow';
 import {STEPS} from './ClassListCreatorPage';
 import available_grade_levels_json from './fixtures/available_grade_levels_json';
