+++ app/assets/javascripts/class_lists/CreateYourLists.story.js
@@ -3,7 +3,7 @@ import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
 import storybookFrame from './storybookFrame';
-import {withDefaultNowContext} from '../../../../spec/javascripts/support/NowContainer';
+import {withDefaultNowContext} from '../testing/NowContainer';
 import CreateYourLists from './CreateYourLists';
 import {
   consistentlyPlacedInitialStudentIdsByRoom,
