+++ app/assets/javascripts/educator/EducatorPage.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {mount} from 'enzyme';
-import {withDefaultNowContext} from '../../../../spec/javascripts/support/NowContainer';
+import {withDefaultNowContext} from '../testing/NowContainer';
 import fetchMock from 'fetch-mock/es5/client';
 import EducatorPage from './EducatorPage';
 import educatorsViewJson from '../../../../spec/javascripts/fixtures/educatorsViewJson';
