+++ app/assets/javascripts/school_courses/SchoolCoursesPage.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
 import SpecSugar from '../../../../spec/javascripts/support/spec_sugar';
-import {withDefaultNowContext} from '../../../../spec/javascripts/support/NowContainer';
+import {withDefaultNowContext} from '../testing/NowContainer';
 import fetchMock from 'fetch-mock/es5/client';
 import SchoolCoursesPage, {SchoolCoursesPagePure} from './SchoolCoursesPage';
 import schoolCoursesJson from '../../../../spec/javascripts/fixtures/schoolCoursesJson';
